Mutualise getAccessRules
