+++ res/css/views/elements/_AddressSelector.scss
@@ -15,9 +15,7 @@ limitations under the License.
 */
 
 .mx_AddressSelector {
-    position: absolute;
     background-color: $primary-bg-color;
-    width: 485px;
     max-height: 116px;
     overflow-y: auto;
 	border-radius: 3px;
